Update URL rendering in feature modal (#44)
* fix: fix ng proxy for geoserver

Geoserver proxy was not set correctly. Data could therefore not be loaded during local development

* refactor: refactor URLs in feature modal

Added separate Ng component for handling URL logic

* fix: fix URL html styling

* feat: make observation goal a clickable URL

Utilize the new app-external-link component

* refactor: use banana box in HTML for better type checking

* feat: prepend http:// if it was not supplied

Check for http(s) at the start of URL. Add http:// if it is not there. URLs will therefore be correctly route, as the url was previously appended to the current url with the absence of http://