Central viewer under some path (#50)
* feat: update files with BASE_HREF in entrypoint

These steps were previously done manually. The index.htmls for each language have to be updated for the new path. Additionally, the nginx config has to be updated with the correct rewrite rule

* fix: assets could no longer be found in the new setup

* refactor: make API_URL more robust by adding leading slash

running central-viewer under path meant that `api` was appended to that. Leading it with a slash (`/api`) will replace the BASE_HREF. This is most probably the place where the API will run anyway